FastAPI/src/main.py
from elasticsearch import AsyncElasticsearch
from fastapi import FastAPI
from fastapi.responses import ORJSONResponse
from redis.asyncio import Redis
from dotenv import load_dotenv
load_dotenv() # load before config file!

from db import elastic, redis
from core import config
from api.v1 import films, genres, persons


app = FastAPI(
    title="Read-only API для онлайн-кинотеатра",
    description="Информация о фильмах, жанрах и людях, участвовавших в создании произведения",
    version="1.0.0",
    docs_url='/api/openapi',
    openapi_url='/api/openapi.json',
    default_response_class=ORJSONResponse,
)


@app.on_event('startup')
async def startup():
    redis.redis = Redis(host=config.REDIS_HOST, port=config.REDIS_PORT)
    elastic.es = AsyncElasticsearch(
        hosts=[f'http://{config.ELASTIC_HOST}:{config.ELASTIC_PORT}'])


@app.on_event('shutdown')
async def shutdown():
    await redis.redis.close()
    await elastic.es.close()

# Подключаем роутер к серверу, указав префикс
app.include_router(films.router, prefix='/api/v1')
app.include_router(persons.router, prefix='/api/v1')
app.include_router(genres.router, prefix='/api/v1')
